Class Field Theory

Class Field Theory

Introduction

This chapter deals with abelian extensions of number fields and the rational numbers.

Class Field Theory, here specifically, class field theory of global number fields, deals with abelian extension, ie. fields where the group of automorphisms is abelian. For extensions of {\Q}, the famous Kronnecker-Weber theorem classifies all such fields: a field is abelian if and only if it is contained in some cyclotomic field. For general number fields this is more involved and even for extensions of {\Q} is is not practical.

In Hecke, abelian extensions are parametrized by quotients of so called ray class groups. The language of ray class groups while dated is more applicable to algorithms than the modern language of idel class groups and quotients.

Ray Class Groups

Given an integral ideal $m_0 \le Z_K$ and a list of real places $m_\infty$, the ray class group modulo $(m_0, m_\infty)$, $C(m)$ is defined as the group of ideals coprime to $m_0$ modulo the elements $a\in K^*$ s.th. $v_p(a-1) \ge v_p(m_0)$ and for all $v\in m_\infty$, $a^{(v)} >0$. This is a finite abelian group. For $m_0 = Z_K$ and $m_\infty = \{\}$ we get $C()$ is the class group, if $m_\infty$ contains all real places, we obtain the narrow class group, or strict class group.

Hecke.ray_class_group — Method.

ray_class_group(m::NfOrdIdl, inf_plc::Array{InfPlc,1}=InfPlc[]; p_part,n_quo)

Given a modulus with finite part $m$ and infinite part $inf_plc$, it returns the Ray Class Group $Cl_m$. If $p_part$ is given, the function will return the largest quotient of the Ray Class Group of p-power order. If $n_quo$ is given, it will return the quotient of the Ray Class Group by n

source

Hecke.class_group — Method.

class_group(O::NfOrd; bound = -1, method = 3, redo = false, large = 1000) -> GrpAbFinGen, Map

Returns a group $A$ and a map $f$ from $A$ to the set of ideals of $O$. The inverse of the map is the projection onto the group of ideals modulo the group of principal ideals. \texttt{redo} allows to trigger a re-computation, thus avoiding the cache. \texttt{bound}, when given, is the bound for the factor base.

source

Hecke.class_group — Method.

class_group(K::AnticNumberField) -> GrpAbFinGen, Map

Shortcut for {{{classgroup(maximalorder(K))}}}: returns the class group as an abelian group and a map from this group to the set of ideals of the maximal order.

source

Hecke.norm_group — Method.

norm_group(f::Nemo.PolyElem, mR::Hecke.MapRayClassGrp, isabelian::Bool = true; of_closure::Bool = false) -> Hecke.FinGenGrpAb, Hecke.FinGenGrpAbMap

norm_group(f::Array{PolyElem{nf_elem}, mR::Hecke.MapRayClassGrp, isabelian::Bool = true; of_closure::Bool = false) -> Hecke.FinGenGrpAb, Hecke.FinGenGrpAbMap

Computes the subgroup of the Ray Class Group $R$ given by the norm of the extension generated by a/the roots of $f$. If {{{isabelian}}} is set to true, then the code assumes the field to be abelian, hence the algorithm stops when the quotient by the norm group has the correct order. Even though the algorithm is probabilistic by nature, in this case the result is guaranteed. If {{{of_closure}}} is given, then the norm group of the splitting field of the polynomial(s) is computed. It is the callers responsibility to ensure that the ray class group passed in is large enough.

source

Hecke.norm_group — Method.

norm_group(K::NfRel{nf_elem}, mR::Hecke.MapRayClassGrp) -> Hecke.FinGenGrpAb, Hecke.FinGenGrpAbMap

norm_group(K::NfRel_ns{nf_elem}, mR::Hecke.MapRayClassGrp) -> Hecke.FinGenGrpAb, Hecke.FinGenGrpAbMap

Computes the subgroup of the Ray Class Group $R$ given by the norm of the extension.

source

Ray Class Fields

In general, the construction of a class field starts with a (ray) class group. Each quotient of a ray class group then defines a ray class field, the defining property is that the (relative) automorphism group is canonically isomorphic to the quotient of the ray class group where the isomorphism is given by the Artin (or Frobenius) map. Since, in Hecke, the (ray) class groups have no link to the field, actually this has to be specified using the maps.

It should be noted that this is a {\em lazy} construction: nothing is computed at this point.

Hecke.ray_class_field — Method.

ray_class_field(m::MapClassGrp) -> ClassField
ray_class_field(m::MapRayClassGrp) -> ClassField

Creates the (formal) abelian extension defined by the map $m: A \to I$ where $I$ is the set of ideals coprime to the modulus defining $m$ and $A$ is a quotient of the ray class group (or class group). The map $m$ must be the map returned from a call to {classgroup} or {rayclass_group}.

source

Hecke.ray_class_field — Method.

ray_class_field(m::Union{MapClassGrp, MapRayClassGrp}, quomap::GrpAbFinGenMap) -> ClassField

For $m$ a map computed by either {rayclassgroup} or {class_group} and $q$ a canonical projection (quotient map) as returned by {quo} for q quotient of the domain of $m$ and a subgroup of $m$, create the (formal) abelian extension where the (relative) automorphism group is canonically isomorphic to the codomain of $q$.

source

Hecke.ray_class_field — Method.

ray_class_field(I::NfAbsOrdIdl; n_quo = 0) -> ClassField

The ray class field modulo $I$. If {{{n_quo}}} is given, then the largest subfield of exponent $n$ is computed.

source

Hecke.hilbert_class_field — Method.

hilbert_class_field(k::AnticNumberField) -> ClassField

The Hilbert class field of $k$ as a formal (ray-) class field.

source

Example

julia> Qx, x = PolynomialRing(FlintQQ, "x");

julia> K, a = NumberField(x^2 - 10, "a");

julia> c, mc = class_group(K)
(Abelian group
Z/2, ClassGroup map of Set of ideals of Maximal order of Number field over Rational Field with defining polynomial x^2-10
with basis nf_elem[1, a]

)

julia> A = ray_class_field(mc)

Conversions

Given a ray class field, it is possible to actually compute defining equation(s) for this field. In general, the number field constructed this way will be non-simple by type and is defined by a polynomial for each maximal cyclic quotient of prime power order in the defining group.

The algorithm employed is based on Kummer-theory and requires the addition of a suitable root of unity. Progress can be monitored by setting {{{setverboselevel(:ClassField, n)}}} where $0\le n\le 3$

Hecke.number_field — Method.

NumberField(f::Generic.Poly{T}, s::String; cached::Bool = false, check::Bool = false) where T

Given an irreducible polynomial $f$ over some number field $K$, create the field $K[t]/f$. $f$ must be irreducible - although this is not tested.

source

NumberField(f::Generic.Poly{T}; cached::Bool = false, check::Bool = false) where T

Given an irreducible polynomial $f$ over some number field $K$, create the field $K[t]/f$. $f$ must be irreducible - although this is not tested.

source

number_field(f::Array{Generic.Poly{T}, 1}, s::String="_\$") where T -> NfRel_ns

Given polynomials $f = (f_1, \ldots, f_n)$ over some number field $k$, construct $K = k[t_1, \ldots, t_n]/\langle f_1(t_1), \ldots, f_n(t_n)\rangle$ The ideal in the quotient must be maximal - although this is not tested.

source

NumberField(CF::ClassField) -> Hecke.NfRel_ns{Nemo.nf_elem}

Given a (formal) abelian extension, compute the class field by finding defining polynomials for all prime power cyclic subfields. Note, by type this is always a non-simple extension.

source

julia> Qx, x = PolynomialRing(FlintQQ, "x");

julia> k, a = NumberField(x^2 - 10, "a");

julia> c, mc = class_group(k);

julia> A = ray_class_field(mc)

julia> K = number_field(A)
non-simple Relative number field over
Number field over Rational Field with defining polynomial x^2-10
 with defining polynomials AbstractAlgebra.Generic.MPoly{nf_elem}[_$1^2+(-2)]

julia> ZK = maximal_order(K)
Relative maximal order of non-simple Relative number field over
Number field over Rational Field with defining polynomial x^2-10
 with defining polynomials AbstractAlgebra.Generic.MPoly{nf_elem}[_$1^2+(-2)]
with pseudo-basis
(1, 1//1 * <1, 1>)
(_$1+(a), 1//4 * <2, a+2>)

julia> isone(discriminant(ZK))
true

Hecke.ray_class_field — Method.

ray_class_field(K::NfRel{nf_elem}) -> ClassField

For a (relative) abelian extension, compute an abstract representation as a class field.

source

Hecke.genus_field — Method.

genus_field(A::ClassField, k::AnticNumberField) -> ClassField

The maximal extension contained in $A$ that is the compositum of $K$ with an abelian extension of $k$.

source

Hecke.maximal_abelian_subfield — Method.

maximal_abelian_subfield(A::ClassField, k::AnticNumberField) -> ClassField

The maximal abelian extension of $k$ contained in $A$. $k$ must be a subfield of the base field of $A$.

source

Hecke.maximal_abelian_subfield — Method.

maximal_abelian_subfield(K::NfRel{nf_elem}; of_closure::Bool = false) -> ClassField

Using a probabilistic algorithm for the norm group computation, determine tha maximal abelian subfield in $K$ over its base field. If {{{of_closure}}} is set to true, then the algorithm is applied to the normal closure if $K$ (without computing it).

source

Invariants

AbstractAlgebra.Generic.degree — Method.

degree(A::ClassField)

The degree of $A$ over its base field, ie. the size of the defining ideal group.

source

AbstractAlgebra.Generic.base_ring — Method.

base_ring(A::ClassField)

The maximal order of the field that $A$ is defined over.

source

Hecke.base_field — Method.

base_field(A::ClassField)

The number field that $A$ is defined over.

source

AbstractAlgebra.Generic.discriminant — Method.

discriminant(C::ClassField) -> NfOrdIdl

Using the conductor-discriminant formula, compute the (relative) discriminant of $C$. This does not use the defining equations.

source

Hecke.conductor — Method.

conductor(C::Hecke.ClassField) -> NfOrdIdl, Array{InfPlc,1}

Return the conductor of the abelian extension corresponding to C

source

Hecke.defining_modulus — Method.

defining_modulus(CF::ClassField)

The modulus, ie. an ideal the the set of real places, used to create the class field.

source

Hecke.iscyclic — Method.

iscyclic(C::ClassField)

Tests if the (relative) automorphism group of $C$ is cyclic (by checking the defining ideal group).

source

Hecke.isconductor — Method.

isconductor(C::Hecke.ClassField, m::NfOrdIdl, inf_plc::Array{InfPlc,1}=InfPlc[]; check) -> NfOrdIdl, Array{InfPlc,1}

Checks if m, inf_plc is the conductor of the abelian extension corresponding to C. If check is false, it assumes that the given modulus is a multiple of the conductor. This is generically faster than computing the conductor.

source

Hecke.isnormal — Method.

isnormal(C::ClassField) -> Bool

For a class field $C$ defined over a normal base field $k$, decide if $C$ is normal over $Q$.

source

Hecke.iscentral — Method.

iscentral(C::ClassField) -> Bool

For a class field $C$ defined over a normal base field $k$, decide if $C$ is central over $Q$.

source

Operations

Base.:* — Method.

*(A::ClassField, B::ClassField) -> ClassField

The compositum of $a$ and $b$ as a (formal) class field.

source

Hecke.compositum — Method.

compositum(a::ClassField, b::ClassField) -> ClassField
         *(a::ClassField, b::ClassField) -> ClassField

The compositum of $a$ and $b$ as a (formal) class field.

source

Base.:== — Method.

==(a::ClassField, b::ClassField)

Tests if $a$ and $b$ are equal.

source

Base.intersect — Method.

intersect(a::ClassField, b::ClassField) -> ClassField

The intersection of $a$ and $b$ as a class field.

source

Hecke.prime_decomposition_type — Method.

prime_decomposition_type(C::ClassField, p::NfAbsOrdIdl) -> (Int, Int, Int)

For a prime $p$ in the base ring of $r$, determine the splitting type of $p$ in $r$. ie. the tuple $(e, f, g)$ giving the ramification degree, the inertia and the number of primes above $p$.

source

Hecke.issubfield — Method.

issubfield(a::ClassField, b::ClassField) -> Bool

Determines of $a$ is a subfield of $b$.

source

Hecke.islocal_norm — Method.

islocal_norm(r::ClassField, a::NfAbsOrdElem) -> Bool

Tests if $a$ is a local norm at all finite places in the extension implictly given by $r$.

source

Hecke.islocal_norm — Method.

islocal_norm(r::ClassField, a::NfAbsOrdElem, p::NfAbsOrdIdl) -> Bool

Tests if $a$ is a local norm at $p$ in the extension implictly given by $r$. Currently the conductor cannot have infinite places.

source

Hecke.normal_closure — Method.

normal_closure(C::ClassField) -> ClassField

For a ray class field $C$ extending a normal base field $k$, compute the normal closure over $Q$.

source

Hecke.subfields — Method.

subfields(C::ClassField) -> Array{ClassField, 1}

Find all subfields of $C$ as class fields. Note: this will not find all subfields over $Q$, but only the ones sharing the same base field.

source

Hecke.subfields — Method.

subfields(C::ClassField, d::Int) -> Array{ClassField, 1}

Find all subfields of $C$ of degree $d$ as class fields. Note: this will not find all subfields over $Q$, but only the ones sharing the same base field.

source