Basics

Basics

Creation and basic properties

Hecke.Order — Method.

Order(B::Array{nf_elem, 1}, check::Bool = true) -> NfOrd

Returns the order with $\mathbf Z$-basis $B$. If check is set, it is checked whether $B$ defines an order.

source

Hecke.Order — Method.

Order(K::AnticNumberField, A::FakeFmpqMat, check::Bool = true) -> NfOrd

Returns the order which has basis matrix $A$ with respect to the power basis of $K$. If check is set, it is checked whether $A$ defines an order.

source

Hecke.Order — Method.

Order(K::AnticNumberField, A::fmpz_mat, check::Bool = true) -> NfOrd

Returns the order which has basis matrix $A$ with respect to the power basis of $K$. If check is set, it is checked whether $A$ defines an order.

source

Order(A::NfOrdFracIdl) -> NfOrd

Returns the fractional ideal $A$ as an order of the ambient number field.

source

  Order(K::RelativeExtension, M::PMat) -> NfRelOrd

Returns the order which has basis pseudo-matrix $M$ with respect to the power basis of $K$.

source

Hecke.EquationOrder — Method.

EquationOrder(K::AnticNumberField) -> NfOrd

Returns the equation order of the number field $K$.

source

Hecke.MaximalOrder — Method.

maximal_order(K::AnticNumberField) -> NfOrd
ring_of_integers(K::AnticNumberField) -> NfOrd

Returns the maximal order of $K$.

Example

julia> Qx, xx = FlintQQ["x"];
julia> K, a = NumberField(x^3 + 2, "a");
julia> O = MaximalOrder(K);

source

Hecke.MaximalOrder — Method.

maximal_order(O::NfOrd) -> NfOrd
MaximalOrder(O::NfOrd) -> NfOrd

Returns the maximal overorder of $O$.

source

Hecke.maximal_order — Method.

maximal_order(K::AnticNumberField) -> NfOrd
ring_of_integers(K::AnticNumberField) -> NfOrd

Returns the maximal order of $K$.

Example

julia> Qx, xx = FlintQQ["x"];
julia> K, a = NumberField(x^3 + 2, "a");
julia> O = MaximalOrder(K);

source

Nemo.lll — Method.

lll(M::NfOrd) -> NfOrd

The same order, but with the basis now being LLL reduced wrt. the Minkowski metric.

source

By Chistov's fundamental theorem, the computation of the maximal order is basically as hard as the factorisation of the discriminant. In order to help the computer, Hecke also provides the following signatures:

Hecke.maximal_order — Method.

MaximalOrder(K::AnticNumberField, primes::Array{fmpz, 1}) -> NfOrd
maximal_order(K::AnticNumberField, primes::Array{fmpz, 1}) -> NfOrd
ring_of_integers(K::AnticNumberField, primes::Array{fmpz, 1}) -> NfOrd

Assuming that $primes$ contains all the prime numbers at which the equation order $\mathbf{Z}[\alpha]$ of $K = \mathbf{Q}(\alpha)$ is not maximal, this function returns the maximal order of $K$.

source

It is also possible the execute the steps individually:

Hecke.pradical — Method.

pradical(O::NfOrd, p::{fmpz|Integer}) -> NfAbsOrdIdl

Given a prime number $p$, this function returns the $p$-radical $\sqrt{p\mathcal O}$ of $\mathcal O$, which is just $\{ x \in \mathcal O \mid \exists k \in \mathbf Z_{\geq 0} \colon x^k \in p\mathcal O \}$. It is not checked that $p$ is prime.

source

Hecke.ring_of_multipliers — Method.

ring_of_multipliers(I::NfAbsOrdIdl) -> NfOrd

Computes the order $(I : I)$, which is the set of all $x \in K$ with $xI \subseteq I$.

source

Example

julia> Qx, x = PolynomialRing(FlintQQ, "x");

julia> K, a = NumberField(x^2 - 2, "a");

julia> O = EquationOrder(K)
Order of Number field over Rational Field with defining polynomial x^2-2
with Z-basis NfAbsOrdElem{AnticNumberField,nf_elem}[1, a]

Base.parent — Method.

parent(O::NfOrd) -> NfOrdSet

Returns the parent of $\mathcal O$, that is, the set of orders of the ambient number field.

source

Hecke.isequation_order — Method.

isequation_order(O::NfOrd) -> Bool

Returns whether $\mathcal O$ is the equation order of the ambient number field.

source

Nemo.signature — Method.

signature(O::NfOrd) -> Tuple{Int, Int}

Returns the signature of the ambient number field of $\mathcal O$.

source

Hecke.nf — Method.

nf(O::NfOrd) -> AnticNumberField

Returns the ambient number field of $\mathcal O$.

source

AbstractAlgebra.Generic.degree — Method.

degree(O::NfOrd) -> Int

Returns the degree of $\mathcal O$.

source

Hecke.basis — Method.

basis(O::NfOrd) -> Array{NfOrdElem, 1}

Returns the $\mathbf Z$-basis of $\mathcal O$.

source

basis(A::NfAbsOrdIdl) -> Array{NfOrdElem, 1}

Returns the basis of A.

source

Hecke.basis — Method.

basis(O::NfOrd, K::AnticNumberField) -> Array{nf_elem, 1}

Returns the $\mathbf Z$-basis elements of $\mathcal O$ as elements of the ambient number field.

source

Hecke.basis_mat — Method.

basis_mat(O::NfOrd) -> FakeFmpqMat

Returns the basis matrix of $\mathcal O$ with respect to the power basis of the ambient number field.

source

basismat(A::NfAbsOrdIdl) -> fmpzmat

Returns the basis matrix of $A$.

source

  basis_mat(O::NfRelOrd{T, S}) -> Generic.Mat{T}

Returns the basis matrix of $\mathcal O$ with respect to the power basis of the ambient number field.

source

  basis_mat(a::NfRelOrdIdl{T, S}) -> Generic.Mat{T}
  basis_mat(a::NfRelOrdFracIdl{T, S}) -> Generic.Mat{T}

Returns the basis matrix of $a$.

source

Hecke.basis_mat_inv — Method.

basis_mat_inv(O::NfOrd) -> FakeFmpqMat

Returns the inverse of the basis matrix of $\mathcal O$.

source

basismatinv(A::NfAbsOrdIdl) -> fmpz_mat

Returns the inverse basis matrix of $A$.

source

  basis_mat_inv(O::NfRelOrd{T, S}) -> Generic.Mat{T}

Returns the inverse of the basis matrix of $\mathcal O$.

source

  basis_mat_inv(a::NfRelOrdIdl{T, S}) -> Generic.Mat{T}
  basis_mat_inv(a::NfRelOrdFracIdl{T, S}) -> Generic.Mat{T}

Returns the inverse of the basis matrix of $a$.

source

AbstractAlgebra.Generic.discriminant — Method.

discriminant(O::NfOrd) -> fmpz

Returns the discriminant of $\mathcal O$.

source

discriminant(B::Array{NfAbsOrdElem, 1}) -> fmpz

Returns the discriminant of the family $B$.

source

Hecke.gen_index — Method.

gen_index(O::NfOrd) -> fmpq

Returns the generalized index of $\mathcal O$ with respect to the equation order of the ambient number field.

source

Hecke.index — Method.

index(O::NfOrd) -> fmpz

Assuming that the order $\mathcal O$ contains the equation order $\mathbf Z[\alpha]$ of the ambient number field, this function returns the index $[ \mathcal O : \mathbf Z]$.

source

Hecke.isindex_divisor — Method.

isindex_divisor(O::NfOrd, d::fmpz) -> Bool
isindex_divisor(O::NfOrd, d::Int) -> Bool

Returns whether $d$ is a divisor of the index of $\mathcal O$. It is assumed that $\mathcal O$ contains the equation order of the ambient number field.

source

Hecke.minkowski_mat — Method.

minkowski_mat(O::NfOrd, abs_tol::Int = 64) -> arb_mat

Returns the Minkowski matrix of $\mathcal O$. Thus if $\mathcal O$ has degree $d$, then the result is a matrix in $\operatorname{Mat}_{d\times d}(\mathbf R)$. The entries of the matrix are real balls of type arb with radius less then 2^-abs_tol.

source

Base.in — Method.

in(a::nf_elem, O::NfOrd) -> Bool

Checks whether $a$ lies in $\mathcal O$.

source

Base.denominator — Method.

denominator(a::nf_elem, O::NfOrd) -> fmpz

Returns the smallest positive integer $k$ such that $k \cdot a$ is contained in $\mathcal O$.

source

Hecke.norm_change_const — Method.

norm_change_const(O::NfOrd) -> (Float64, Float64)

Returns $(c_1, c_2) \in \mathbf R_{>0}^2$ such that for all $x = \sum_{i=1}^d x_i \omega_i \in \mathcal O$ we have $T_2(x) \leq c_1 \cdot \sum_i^d x_i^2$ and $\sum_i^d x_i^2 \leq c_2 \cdot T_2(x)$, where $(\omega_i)_i$ is the $\mathbf Z$-basis of $\mathcal O$.

source

Hecke.trace_matrix — Method.

trace_matrix(O::NfOrd) -> fmpz_mat

Returns the trace matrix of \mathcal O, that is, the matrix $(\operatorname{tr}_{K/\mathbf Q}(b_i \cdot b_j))_{1 \leq i, j \leq d}$.

source

Base.:+ — Method.

+(R::NfOrd, S::NfOrd) -> NfOrd

Given two orders $R$, $S$ of $K$, this function returns the smallest order containing both $R$ and $S$. It is assumed that $R$, $S$ contain the ambient equation order and have coprime index.

source

Hecke.poverorder — Method.

poverorder(O::NfOrd, p::fmpz) -> NfOrd
poverorder(O::NfOrd, p::Integer) -> NfOrd

This function tries to find an order that is locally larger than $\mathcal O$ at the prime $p$: If $p$ divides the index $[ \mathcal O_K : \mathcal O]$, this function will return an order $R$ such that $v_p([ \mathcal O_K : R]) < v_p([ \mathcal O_K : \mathcal O])$. Otherwise $\mathcal O$ is returned.

source

Hecke.pmaximal_overorder — Method.

pmaximal_overorder(O::NfOrd, p::fmpz) -> NfOrd
pmaximal_overorder(O::NfOrd, p::Integer) -> NfOrd

This function finds a $p$-maximal order $R$ containing $\mathcal O$. That is, the index $[ \mathcal O_K : R]$ is not dividible by $p$.

source

Elements

Creation

Basic properties

Base.parent — Method.

parent(a::NfAbsOrdElem) -> NfOrd

Returns the order of which $a$ is an element.

source

Hecke.elem_in_nf — Method.

elem_in_nf(a::NfAbsOrdElem) -> nf_elem

Returns the element $a$ considered as an element of the ambient number field.

source

Hecke.elem_in_basis — Method.

elem_in_basis(a::NfAbsOrdElem) -> Array{fmpz, 1}

Returns the coefficient vector of $a$.

source

  elem_in_basis(a::NfRelOrdElem{T}) -> Vector{T}

Returns the coefficient vector of $a$.

source

AbstractAlgebra.Generic.discriminant — Method.

discriminant(B::Array{NfAbsOrdElem, 1}) -> fmpz

Returns the discriminant of the family $B$.

source

Base.:== — Method.

==(x::NfAbsOrdElem, y::NfAbsOrdElem) -> Bool

Returns whether $x$ and $y$ are equal.

source

Base.zero — Method.

zero(O::NfOrd) -> NfAbsOrdElem

Returns the zero element of $\mathcal O$.

source

Base.one — Method.

one(O::NfOrd) -> NfAbsOrdElem

Returns the one element of $\mathcal O$.

source

Base.iszero — Method.

iszero(a::NfOrd) -> Bool

Tests if $a$ is zero.

source

Base.isone — Method.

isone(a::NfOrd) -> Bool

Tests if $a$ is one.

source

Arithmetic

Base.:- — Method.

-(x::NfAbsOrdElem) -> NfAbsOrdElem

Returns the additive inverse of $x$.

source

Base.:+ — Method.

+(x::NfAbsOrdElem, y::NfAbsOrdElem) -> NfAbsOrdElem

Returns $x + y$.

source

Base.:- — Method.

-(x::NfAbsOrdElem, y::NfAbsOrdElem) -> NfAbsOrdElem

Returns $x - y$.

source

Base.:* — Method.

*(x::NfAbsOrdElem, y::NfAbsOrdElem) -> NfAbsOrdElem

Returns $x \cdot y$.

source

Base.:^ — Method.

^(x::NfAbsOrdElem, y::Union{fmpz, Int})

Returns $x^y$.

source

Base.mod — Method.

mod(a::NfAbsOrdElem, m::Union{fmpz, Int}) -> NfAbsOrdElem

Reduces the coefficient vector of $a$ modulo $m$ and returns the corresponding element. The coefficient vector of the result will have entries $x$ with $0 \leq x \leq m$.

source

Base.powermod — Method.

powermod(a::NfAbsOrdElem, i::fmpz, m::Integer) -> NfAbsOrdElem

Returns the element $a^i$ modulo $m$.

source

Miscallenous

Nemo.representation_matrix — Method.

representation_matrix(a::NfAbsOrdElem) -> fmpz_mat

Returns the representation matrix of the element $a$.

source

representation_matrix(a::NfAbsOrdElem, K::AnticNumberField) -> FakeFmpqMat

Returns the representation matrix of the element $a$ considered as an element of the ambient number field $K$. It is assumed that $K$ is the ambient number field of the order of $a$.

source

Nemo.representation_matrix — Method.

representation_matrix(a::NfAbsOrdElem, K::AnticNumberField) -> FakeFmpqMat

Returns the representation matrix of the element $a$ considered as an element of the ambient number field $K$. It is assumed that $K$ is the ambient number field of the order of $a$.

source

LinearAlgebra.tr — Method.

tr(a::NfAbsOrdElem) -> fmpz

Returns the trace of $a$.

source

LinearAlgebra.norm — Method.

norm(a::NfAbsOrdElem) -> fmpz

Returns the norm of $a$.

source

norm(A::NfAbsOrdIdl) -> fmpz

Returns the norm of $A$, that is, the cardinality of $\mathcal O/A$, where $\mathcal O$ is the order of $A$.

source

norm(a::NfRelOrdIdl) -> NfOrdIdl

Returns the norm of $a$.

source

norm(a::NfRelOrdFracIdl{T, S}) -> S

Returns the norm of $a$

source

Base.rand — Method.

rand(O::NfOrd, n::Union{Integer, fmpz}) -> NfAbsOrdElem

Computes a coefficient vector with entries uniformly distributed in $\{-n,\dotsc,-1,0,1,\dotsc,n\}$ and returns the corresponding element of the order $\mathcal O$.

source

Hecke.minkowski_map — Method.

minkowski_map(a::NfAbsOrdElem, abs_tol::Int) -> Array{arb, 1}

Returns the image of $a$ under the Minkowski embedding. Every entry of the array returned is of type arb with radius less then 2^-abs_tol.

source

Hecke.conjugates_arb — Method.

conjugates_arb(x::NfAbsOrdElem, abs_tol::Int) -> Array{acb, 1}

Compute the the conjugates of x as elements of type acb. Recall that we order the complex conjugates $\sigma_{r+1}(x),...,\sigma_{r+2s}(x)$ such that $\sigma_{i}(x) = \overline{\sigma_{i + s}(x)}$ for $r + 2 \leq i \leq r + s$.

Every entry y of the array returned satisfies radius(real(y)) < 2^-abs_tol, radius(imag(y)) < 2^-abs_tol respectively.

source

Hecke.conjugates_arb_log — Method.

conjugates_arb_log(x::NfAbsOrdElem, abs_tol::Int) -> Array{arb, 1}

Returns the elements $(\log(\lvert \sigma_1(x) \rvert),\dotsc,\log(\lvert\sigma_r(x) \rvert), \dotsc,2\log(\lvert \sigma_{r+1}(x) \rvert),\dotsc, 2\log(\lvert \sigma_{r+s}(x)\rvert))$ as elements of type arb radius less then 2^-abs_tol.

source

Hecke.t2 — Method.

t2(x::NfAbsOrdElem, abs_tol::Int = 32) -> arb

Return the $T_2$-norm of $x$. The radius of the result will be less than 2^-abs_tol.

source

AbstractAlgebra.Generic.minpoly — Method.

minpoly(a::NfAbsOrdElem) -> fmpz_poly

minpoly(a::NfAbsOrdElem, FlintZZ) -> fmpz_poly

The minimal polynomial of $a$.

source

AbstractAlgebra.Generic.charpoly — Method.

charpoly(a::NfAbsOrdElem) -> fmpz_poly

charpoly(a::NfAbsOrdElem, FlintZZ) -> fmpz_poly

The characteristic polynomial of $a$.

source

Ideals

Creation

Hecke.ideal — Method.

ideal(O::NfOrd, a::fmpz) -> NfAbsOrdIdl
ideal(O::NfOrd, a::Integer) -> NfAbsOrdIdl

Returns the ideal of $\mathcal O$ which is generated by $a$.

source

Hecke.ideal — Method.

ideal(O::NfOrd, x::fmpz_mat, check::Bool = false, x_in_hnf::Bool = false) -> NfAbsOrdIdl

Creates the ideal of $\mathcal O$ with basis matrix $x$. If check is set, then it is checked whether $x$ defines an ideal (expensive). If xinhnf is set, then it is assumed that $x$ is already in lower left HNF.

source

Hecke.ideal — Method.

ideal(O::NfOrd, x::NfOrdElem) -> NfAbsOrdIdl

Creates the principal ideal $(x)$ of $\mathcal O$.

source

Hecke.ideal — Method.

ideal(O::NfOrd, x::fmpz, y::NfOrdElem) -> NfAbsOrdIdl
ideal(O::NfOrd, x::Integer, y::NfOrdElem) -> NfAbsOrdIdl

Creates the ideal $(x,y)$ of $\mathcal O$.

source

Base.:* — Method.

*(O::NfOrd, x::NfOrdElem) -> NfAbsOrdIdl
*(x::NfOrdElem, O::NfOrd) -> NfAbsOrdIdl

Returns the principal ideal $(x)$ of $\mathcal O$.

source

Hecke.prime_decomposition — Method.

prime_decomposition(O::NfOrd,
                    p::Integer,
                    degree_limit::Int = 0,
                    lower_limit::Int = 0) -> Array{Tuple{NfOrdIdl, Int}, 1}

Returns an array of tuples $(\mathfrak p_i,e_i)$ such that $p \mathcal O$ is the product of the $\mathfrak p_i^{e_i}$ and $\mathfrak p_i \neq \mathfrak p_j$ for $i \neq j$.

If degree_limit is a nonzero integer $k > 0$, then only those prime ideals $\mathfrak p$ with $\deg(\mathfrak p) \leq k$ will be returned. Similarly if \lower_limit is a nonzero integer $l > 0$, then only those prime ideals $\mathfrak p$ with $l \leq \deg(\mathfrak p)$ will be returned. Note that in this case it may happen that $p\mathcal O$ is not the product of the $\mathfrak p_i^{e_i}$.

source

Hecke.prime_decomposition — Method.

prime_decomposition(O::NfOrd,
                    p::Integer,
                    degree_limit::Int = 0,
                    lower_limit::Int = 0) -> Array{Tuple{NfOrdIdl, Int}, 1}

Returns an array of tuples $(\mathfrak p_i,e_i)$ such that $p \mathcal O$ is the product of the $\mathfrak p_i^{e_i}$ and $\mathfrak p_i \neq \mathfrak p_j$ for $i \neq j$.

If degree_limit is a nonzero integer $k > 0$, then only those prime ideals $\mathfrak p$ with $\deg(\mathfrak p) \leq k$ will be returned. Similarly if \lower_limit is a nonzero integer $l > 0$, then only those prime ideals $\mathfrak p$ with $l \leq \deg(\mathfrak p)$ will be returned. Note that in this case it may happen that $p\mathcal O$ is not the product of the $\mathfrak p_i^{e_i}$.

source

Arithmetic

Base.:== — Method.

==(x::NfAbsOrdIdl, y::NfAbsOrdIdl)

Returns whether $x$ and $y$ are equal.

source

Base.:+ — Method.

+(x::NfOrdIdl, y::NfOrdIdl)

Returns $x + y$.

source

Base.:* — Method.

*(x::NfOrdIdl, y::NfOrdIdl)

Returns $x \cdot y$.

source

Base.lcm — Method.

intersection(x::NfOrdIdl, y::NfOrdIdl) -> NfOrdIdl
lcm(x::NfOrdIdl, y::NfOrdIdl) -> NfOrdIdl

Returns $x \cap y$.

source

Miscaellenous

AbstractAlgebra.Generic.order — Method.

order(I::NfAbsOrdIdl) -> NfOrd

Returns the order of $I$.

source

Hecke.basis — Method.

basis(O::NfOrd) -> Array{NfOrdElem, 1}

Returns the $\mathbf Z$-basis of $\mathcal O$.

source

basis(A::NfAbsOrdIdl) -> Array{NfOrdElem, 1}

Returns the basis of A.

source

Hecke.basis_mat — Method.

basis_mat(O::NfOrd) -> FakeFmpqMat

Returns the basis matrix of $\mathcal O$ with respect to the power basis of the ambient number field.

source

basismat(A::NfAbsOrdIdl) -> fmpzmat

Returns the basis matrix of $A$.

source

  basis_mat(O::NfRelOrd{T, S}) -> Generic.Mat{T}

Returns the basis matrix of $\mathcal O$ with respect to the power basis of the ambient number field.

source

  basis_mat(a::NfRelOrdIdl{T, S}) -> Generic.Mat{T}
  basis_mat(a::NfRelOrdFracIdl{T, S}) -> Generic.Mat{T}

Returns the basis matrix of $a$.

source

Hecke.basis_mat_inv — Method.

basis_mat_inv(O::NfOrd) -> FakeFmpqMat

Returns the inverse of the basis matrix of $\mathcal O$.

source

basismatinv(A::NfAbsOrdIdl) -> fmpz_mat

Returns the inverse basis matrix of $A$.

source

  basis_mat_inv(O::NfRelOrd{T, S}) -> Generic.Mat{T}

Returns the inverse of the basis matrix of $\mathcal O$.

source

  basis_mat_inv(a::NfRelOrdIdl{T, S}) -> Generic.Mat{T}
  basis_mat_inv(a::NfRelOrdFracIdl{T, S}) -> Generic.Mat{T}

Returns the inverse of the basis matrix of $a$.

source

Base.minimum — Method.

minimum(A::NfAbsOrdIdl) -> fmpz

Returns the smallest nonnegative element in $A \cap \mathbf Z$.

source

  minimum(A::NfRelOrdIdl) -> NfOrdIdl
  minimum(A::NfRelOrdIdl) -> NfRelOrdIdl

Returns the ideal $A \cap O$ where $O$ is the maximal order of the coefficient ideals of $A$.

source

LinearAlgebra.norm — Method.

norm(A::NfAbsOrdIdl) -> fmpz

Returns the norm of $A$, that is, the cardinality of $\mathcal O/A$, where $\mathcal O$ is the order of $A$.

source

norm(a::NfRelOrdIdl) -> NfOrdIdl

Returns the norm of $a$.

source

norm(a::NfRelOrdFracIdl{T, S}) -> S

Returns the norm of $a$

source

Base.in — Method.

in(x::NfOrdElem, y::NfAbsOrdIdl)
in(x::nf_elem, y::NfAbsOrdIdl)
in(x::fmpz, y::NfAbsOrdIdl)

Returns whether $x$ is contained in $y$.

source

Hecke.idempotents — Method.

idempotents(x::NfOrdIdl, y::NfOrdIdl) -> NfOrdElem, NfOrdElem

Returns a tuple (e, f) consisting of elements e in x, f in y such that 1 = e + f.

If the ideals are not coprime, an error is raised.

source

Base.mod — Method.

mod(x::NfOrdElem, I::NfAbsOrdIdl)

Returns the unique element $y$ of the ambient order of $x$ with $x \equiv y \bmod I$ and the following property: If $a_1,\dotsc,a_d \in \Z_{\geq 1}$ are the diagonal entries of the unique HNF basis matrix of $I$ and $(b_1,\dotsc,b_d)$ is the coefficient vector of $y$, then $0 \leq b_i < a_i$ for $1 \leq i \leq d$.

source

Nemo.isprime — Method.

isprime(A::NfOrdIdl) -> Bool

Returns whether $A$ is a prime ideal.

source

AbstractAlgebra.Generic.valuation — Method.

valuation(a::nf_elem, p::NfOrdIdl) -> fmpz
valuation(a::NfOrdElem, p::NfOrdIdl) -> fmpz
valuation(a::fmpz, p::NfOrdIdl) -> fmpz

Computes the $\mathfrak p$-adic valuation of $a$, that is, the largest $i$ such that $a$ is contained in $\mathfrak p^i$.

source

AbstractAlgebra.Generic.valuation — Method.

valuation(a::nf_elem, p::NfOrdIdl) -> fmpz
valuation(a::NfOrdElem, p::NfOrdIdl) -> fmpz
valuation(a::fmpz, p::NfOrdIdl) -> fmpz

Computes the $\mathfrak p$-adic valuation of $a$, that is, the largest $i$ such that $a$ is contained in $\mathfrak p^i$.

source

AbstractAlgebra.Generic.valuation — Method.

valuation(A::NfOrdIdl, p::NfOrdIdl) -> fmpz

Computes the $\mathfrak p$-adic valuation of $A$, that is, the largest $i$ such that $A$ is contained in $\mathfrak p^i$.

source

AbstractAlgebra.Generic.valuation — Method.

valuation(a::Integer, p::NfOrdIdl) -> fmpz

Computes the $\mathfrak p$-adic valuation of $a$, that is, the largest $i$ such that $a$ is contained in $\mathfrak p^i$.

source

AbstractAlgebra.Generic.valuation — Method.

valuation(a::nf_elem, p::NfOrdIdl) -> fmpz
valuation(a::NfOrdElem, p::NfOrdIdl) -> fmpz
valuation(a::fmpz, p::NfOrdIdl) -> fmpz

Computes the $\mathfrak p$-adic valuation of $a$, that is, the largest $i$ such that $a$ is contained in $\mathfrak p^i$.

source

AbstractAlgebra.Generic.valuation — Method.

valuation(A::NfOrdFracIdl, p::NfOrdIdl)

The valuation of $A$ at $p$.

source

Fractional ideals

Creation

Hecke.frac_ideal — Method.

frac_ideal(O::NfOrd, A::fmpz_mat, b::fmpz, A_in_hnf::Bool = false) -> NfOrdFracIdl

Creates the fractional ideal of $\mathcal O$ with basis matrix $A/b$. If Ainhnf is set, then it is assumed that $A$ is already in lower left HNF.

source

Hecke.frac_ideal — Method.

frac_ideal(O::NfOrd, A::fmpz_mat, b::fmpz, A_in_hnf::Bool = false) -> NfOrdFracIdl

Creates the fractional ideal of $\mathcal O$ with basis matrix $A/b$. If Ainhnf is set, then it is assumed that $A$ is already in lower left HNF.

source

Hecke.frac_ideal — Method.

frac_ideal(O::NfOrd, A::FakeFmpqMat, A_in_hnf::Bool = false) -> NfOrdFracIdl

Creates the fractional ideal of $\mathcal O$ with basis matrix $A$. If Ainhnf is set, then it is assumed that the numerator of $A$ is already in lower left HNF.

source

Hecke.frac_ideal — Method.

frac_ideal(O::NfOrd, I::NfOrdIdl) -> NfOrdFracIdl

Turns the ideal $I$ into a fractional ideal of $\mathcal O$.

source

Hecke.frac_ideal — Method.

frac_ideal(O::NfOrd, I::NfOrdIdl, b::fmpz) -> NfOrdFracIdl

Creates the fractional ideal $I/b$ of $\mathcal O$.

source

Hecke.frac_ideal — Method.

frac_ideal(O::NfOrd, a::nf_elem) -> NfOrdFracIdl

Creates the principal fractional ideal $(a)$ of $\mathcal O$.

source

Hecke.frac_ideal — Method.

frac_ideal(O::NfOrd, a::NfOrdElem) -> NfOrdFracIdl

Creates the principal fractional ideal $(a)$ of $\mathcal O$.

source

Arithmetic

Base.:== — Method.

==(x::NfOrdFracIdl, y::NfOrdFracIdl) -> Bool

Returns whether $x$ and $y$ are equal.

source

Base.inv — Method.

inv(A::NfOrdFracIdl) -> NfOrdFracIdl

Returns the fractional ideal $B$ such that $AB = \mathcal O$.

source

Hecke.integral_split — Method.

integral_split(A::NfOrdFracIdl) -> NfOrdIdl, NfOrdIdl

Computes the unique coprime integral ideals $N$ and $D$ s.th. $A = ND^{-1}$

source

Miscaellenous

AbstractAlgebra.Generic.order — Method.

order(a::NfOrdFracIdl) -> NfOrd

The order that was used to define the ideal $a$.

source

Hecke.basis_mat — Method.

basis_mat(I::NfOrdFracIdl) -> FakeFmpqMat

Returns the basis matrix of $I$ with respect to the basis of the order.

source

Hecke.basis_mat_inv — Method.

basis_mat_inv(I::NfOrdFracIdl) -> FakeFmpqMat

Returns the inverse of the basis matrix of $I$.

source

Hecke.basis — Method.

basis(I::NfOrdFracIdl) -> Array{nf_elem, 1}

Returns the $\mathbf Z$-basis of $I$.

source

LinearAlgebra.norm — Method.

norm(I::NfOrdFracIdl) -> fmpq

Returns the norm of $I$

source